chore(eslint): follow eslint v9 plugin specification
The plugin should be the default export, configuration is exported as recommended.
See https://eslint.org/docs/latest/extend/plugins#create-a-plugin

eslint-config export is kept for backward compatibility
